Feature/202304023

+++ cmd/server/v1/domain/endpoint.go
@@ -8,7 +8,7 @@ import (
 type EndpointReqPaginate struct {
 	_domain.PaginateReq
 	ProjectId    int64  `json:"projectId"`
-	CategoryId   uint   `json:"categoryId"`
+	CategoryId   int64  `json:"categoryId"`
 	Status       int64  `json:"status"`
 	CreateUser   string `json:"createUser"`
 	Title        string `json:"title"`
@@ -48,7 +48,7 @@ type EndpointReq struct {
 	Path        string          `json:"path"`
 	Version     string          `json:"version"`
 	CreateUser  string          `json:"CreateUser"`
-	CategoryId  uint            `json:"categoryId"`
+	CategoryId  int64           `json:"categoryId"`
 	PathParams  []OpenApiParam  `gorm:"-" json:"pathParams"`
 	Interfaces  []InterfaceResp `gorm:"-" json:"interfaces"`
 	Description string          `json:"description"`
